% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anthroBoot}
\alias{anthroBoot}
\title{Boostrap results for anthropometric indicators}
\format{A data.frame with 20 rows and 6 columns
\describe{
  \item{\code{varLabel}}{Short code of indicators}
  \item{\code{df}}{Name of data.frame to which indicator belongs}
  \item{\code{varNames}}{Long names of indicators}
  \item{\code{estimate}}{Indicator estimate}
  \item{\code{lcl}}{Indiator 95\% lower confidence limits}
  \item{\code{ucl}}{Indiator 95\% lower confidence limits}
}}
\usage{
anthroBoot
}
\description{
Boostrap results for anthropometric indicators
}
\keyword{datasets}
