% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-boot_indicators.R
\name{boot_estimate}
\alias{boot_estimate}
\title{Apply blocked weighted bootstrap to indicator datasets}
\usage{
boot_estimate(indicator = c("ifaDF", "iycfDF", "mnpDF", "vitDF",
  "screenDF", "anthroDF"), county, w, replicates = 399,
  core.columns = c("spid", "cid", "did", "eid", "motherID", "m2"))
}
\arguments{
\item{indicator}{A character vector of indicator data.frame names}

\item{county}{An integer indicating which county to interpolate; 1 for
Greater Monrovia; 2 for Grand Bassa}

\item{w}{Population dataset}

\item{replicates}{Number of bootstrap replicates}

\item{core.columns}{A vector of variable names included in indicator
data.frames}
}
\value{
A data.frame of indicator estimates with lower and upper confidence
  limits
}
\description{
A wrapper for \code{bootBW} function in \code{bbw} package.
}
\examples{
boot_estimate(indicator = "iycfDF", county = 1, w = psuDataGM, replicates = 9)

}
