% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmamDF}
\alias{cmamDF}
\title{Community-based Management of Acute Malnutrition (CMAM) coverage indicators
Round 2}
\format{A data.frame with 35 columns:
\describe{
  \item{\code{spid}}{Sampling point identifier}
  \item{\code{cid}}{County identifier. 1 = Urban Montserrado; 2 = Grand Bassa}
  \item{\code{did}}{District identifier}
  \item{\code{eid}}{Enumeration area identifier}
  \item{\code{motherID}}{Mother identifier}
  \item{\code{m2}}{Mother age}
  \item{\code{cin}}{SAM case in programme}
  \item{\code{cout}}{SAM case not in programme}
  \item{\code{rin}}{Recovering case in programme}
  \item{\code{q1}}{Is child sick?}
  \item{\code{q1a}}{Child suffering from: malnutrition}
  \item{\code{q1b}}{Child suffering from: diarrhoea}
  \item{\code{q1c}}{Child suffering from: fever}
  \item{\code{q1d}}{Child suffering from: cough}
  \item{\code{q1e}}{Child suffering from: malaria}
  \item{\code{q1f}}{Child suffering from: don't know}
  \item{\code{q2}}{Do you think your child is too small?}
  \item{\code{q3}}{Know of treatment}
  \item{\code{q4a}}{Came to health centre but was not measured}
  \item{\code{q4b}}{Too far}
  \item{\code{q4c}}{No time/too busy}
  \item{\code{q4d}}{Mother/carer is sick}
  \item{\code{q4e}}{Cannot carry more than one child}
  \item{\code{q4f}}{Feels ashamed about coming}
  \item{\code{q4g}}{Security problems}
  \item{\code{q4h}}{No one else to take care of other siblings}
  \item{\code{q4i}}{Rejected by programme}
  \item{\code{q4j}}{Other parent's children have been rejected}
  \item{\code{q4k}}{Husband refused}
  \item{\code{q4l}}{Thought necessary to be enrolled in hospital first}
  \item{\code{q4m}}{Does not think programme can help children}
  \item{\code{q4n}}{Got diarrhoea}
  \item{\code{q4o}}{Ration has worms}
  \item{\code{q4p}}{Out of stock}
  \item{\code{q4q}}{Lack of money to pay for transport}
  \item{\code{q4r}}{SAM case in SFP}
}}
\source{
Liberia Coverage Survey Round 2 - 27 August to 30 September 2019
}
\usage{
cmamDF
}
\description{
Community-based Management of Acute Malnutrition (CMAM) coverage indicators
Round 2
}
\keyword{datasets}
