% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmamEst}
\alias{cmamEst}
\title{Boostrap results for CMAM coverage indicators}
\format{
A data.frame with 2 rows and 7 columns
\describe{
  \item{\code{county}}{County name}
  \item{\code{varLabel}}{Short code of indicators}
  \item{\code{df}}{Name of data.frame to which indicator belongs}
  \item{\code{varNames}}{Long names of indicators}
  \item{\code{estimate}}{Indicator estimate}
  \item{\code{lcl}}{Indiator 95\% lower confidence limits}
  \item{\code{ucl}}{Indiator 95\% lower confidence limits}
}
}
\usage{
cmamEst
}
\description{
Boostrap results for CMAM coverage indicators
}
\keyword{datasets}
