% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmamInt}
\alias{cmamInt}
\title{Interpolated Community-based Management of Acute Malnutrition (CMAM) coverage
indicator set Round 2}
\format{A data.frame with 29 columns:
\describe{
  \item{\code{q1}}{Is child sick?}
  \item{\code{q1a}}{Child suffering from: malnutrition}
  \item{\code{q1b}}{Child suffering from: diarrhoea}
  \item{\code{q1c}}{Child suffering from: fever}
  \item{\code{q1d}}{Child suffering from: cough}
  \item{\code{q1e}}{Child suffering from: malaria}
  \item{\code{q1f}}{Child suffering from: don't know}
  \item{\code{q2}}{Do you think your child is too small?}
  \item{\code{q3}}{Know of treatment}
  \item{\code{q4a}}{Came to health centre but was not measured}
  \item{\code{q4b}}{Too far}
  \item{\code{q4c}}{No time/too busy}
  \item{\code{q4d}}{Mother/carer is sick}
  \item{\code{q4e}}{Cannot carry more than one child}
  \item{\code{q4f}}{Feels ashamed about coming}
  \item{\code{q4g}}{Security problems}
  \item{\code{q4h}}{No one else to take care of other siblings}
  \item{\code{q4i}}{Rejected by programme}
  \item{\code{q4j}}{Other parent's children have been rejected}
  \item{\code{q4k}}{Husband refused}
  \item{\code{q4l}}{Thought necessary to be enrolled in hospital first}
  \item{\code{q4m}}{Does not think programme can help children}
  \item{\code{q4n}}{Got diarrhoea}
  \item{\code{q4o}}{Ration has worms}
  \item{\code{q4p}}{Out of stock}
  \item{\code{q4q}}{Lack of money to pay for transport}
  \item{\code{q4r}}{SAM case in SFP}
  \item{\code{cf}}{Case-finding effectiveness}
  \item{\code{tc}}{Treatment coverage}
}}
\source{
Liberia Coverage Survey Round 2 - 27 August to 30 September 2019
}
\usage{
cmamInt
}
\description{
Interpolated Community-based Management of Acute Malnutrition (CMAM) coverage
indicator set Round 2
}
\keyword{datasets}
