% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coverageData1.r2}
\alias{coverageData1.r2}
\title{Liberia Coverage Survey Data Round 2 - main form}
\format{A data.frame with 69 columns
\describe{
  \item{\code{sdate}}{Date of data collection}
  \item{\code{cid}}{County identifier}
  \item{\code{county}}{County name}
  \item{\code{did}}{District identifier}
}}
\source{
Liberia Coverage Survey Round 2 - 27 August to 30 September 2019
}
\usage{
coverageData1.r2
}
\description{
Liberia Coverage Survey Data Round 2 - main form
}
\keyword{datasets}
