% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coverageData2.r2}
\alias{coverageData2.r2}
\title{Liberia Coverage Survey Data Round 2 - repeat form}
\format{A data.frame with 65 columns

* `sdate` - Date of data collection
* `cid` - County identifier
* `county` - County name
* `did` - District identifier}
\source{
Liberia Coverage Survey Round 2 - 27 August to 30 September 2019
}
\usage{
coverageData2.r2
}
\description{
Liberia Coverage Survey Data Round 2 - repeat form
}
\keyword{datasets}
