% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-create_id.R
\name{create_id}
\alias{create_id}
\title{Function to create unique mother and child identifiers}
\usage{
create_id(ls)
}
\arguments{
\item{ls}{A list of data.frames produced when retrieving the Liberia Coverage
Survey data from the ODK server}
}
\value{
A list of two vectors each with length equal to the number of rows of
  the corresponding data.frame in `ls`
}
\description{
Function to create unique mother and child identifiers
}
\examples{
\dontrun{
create_id(ls = coverageData)
}

}
