% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-create_sp.R
\name{create_points}
\alias{create_points}
\title{Create SpatialPoints from a hexagonal grid map}
\usage{
create_points(hexgrid)
}
\arguments{
\item{hexgrid}{A SpatialPolygonsDataFrame hexagonal grid map}
}
\value{
A SpatialPoints class object of hexagonal grid centroids
}
\description{
Create SpatialPoints from a hexagonal grid map
}
\examples{
create_points(hexgrid = gmHexGrid)

}
