% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-estimate_classic.R
\name{estimate_classic}
\alias{estimate_classic}
\title{Estimate proportion using classic estimator}
\usage{
estimate_classic(x, n, z = 1.96)
}
\arguments{
\item{x}{Total number of successes}

\item{n}{Total number of tests}

\item{z}{z-value. Default is 1.96 for 95\% confidence interval}
}
\value{
A list containing the following components:
\item{\code{estimate}}{A vector with the sample proportions}
\item{\code{conf.int}}{A confidence interval for the true proportion}
}
\description{
Estimate proportion using classic estimator
}
\examples{

estimate_classic(x = 50, n = 200)

}
