% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gbHexGrid}
\alias{gbHexGrid}
\title{Grand Bassa hexagonal grid map}
\format{A SpatialPolygonsDataFrame of 8670 hexagons across Grand Bassa}
\usage{
gbHexGrid
}
\description{
Grand Bassa hexagonal grid map
}
\keyword{datasets}
