% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-get_data.R
\name{get_liberia_data}
\alias{get_liberia_data}
\title{Function to get Liberia Coverage Survey datasets and forms from ODK server.
This function is a wrapper for a set of functions from the \code{odkr}
package.}
\usage{
get_liberia_data(id, username, password, start = Sys.Date(),
  end = Sys.Date(), filename)
}
\arguments{
\item{id}{Form ID. Can be one of two form IDs used by the
different forms designed for the Liberia Coverage Survey -
\code{liberiaCoverageRound2} and \code{liberiaVillageFormRound2}.}

\item{username}{ONA server username credentials. For}

\item{password}{ONA server password credentials.}

\item{start}{Include data from submission dates after (inclusive) this
start date in export to CSV. Date format <yyyy/MM/dd> and default is
current system date}

\item{end}{Include data from submission dates before (exclusive) this date
in export to CSV. Date format is <yyyy/MM/dd> and default value current
system date}

\item{filename}{Filename to use for data without the CSV file extension.}
}
\value{
A data.frame corresponding to dataset corresponding to the form ID
  specified.
}
\description{
Function to get Liberia Coverage Survey datasets and forms from ODK server.
This function is a wrapper for a set of functions from the \code{odkr}
package.
}
\examples{
\dontrun{
  get_liberia_data(id = "liberiaCoverageRound2",
                   username = "cadnihead",
                   password = "kEv-hAB-Arb-6Cn",
                   filename = "surveyData")
}

}
