% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-get_spid.R
\name{get_spid}
\alias{get_spid}
\title{Function to get sampling point identifiers for Liberia Coverage dataset}
\usage{
get_spid(df, splist)
}
\arguments{
\item{df}{A coverage data.frame collected for the Liberia Coverage Survey}

\item{splist}{A data.frame containing sampling point identifier information
per enumeration area}
}
\value{
Coverage data.frame with corresponding sampling point identifier
  per enumeration area
}
\description{
Function to get sampling point identifiers for Liberia Coverage dataset
}
\examples{
get_spid(df = coverageData.r2, splist = sampleList.r2[ , c("spid", "EFEACODE")])

}
