% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gmHexGrid}
\alias{gmHexGrid}
\title{Greater Monrovia hexagonal grid map}
\format{A SpatialPolygonsDataFrame of 8663 hexagons across Monrovia}
\usage{
gmHexGrid
}
\description{
Greater Monrovia hexagonal grid map
}
\keyword{datasets}
