% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ifaInt}
\alias{ifaInt}
\title{Interpolated iron-folic acid (IFA) indicator set Round 2}
\format{A list of 2 data.frames each with 20 columns:
\describe{
  \item{\code{ifa1}}{Go to antenatal care?}
  \item{\code{ifa2}}{Received information about IFA?}
  \item{\code{ifa3}}{Received/purchased IFA?}
  \item{\code{ifa3a}}{Reasons: Health centre ran out}
  \item{\code{ifa3b}}{Reasons: Took too long to get tablets}
  \item{\code{ifa3c}}{Too expensive}
  \item{\code{ifa3d}}{Did not go to hospital}
  \item{\code{ifa3e}}{Don't know about the tablets}
  \item{\code{ifa3f}}{Didn't get any tablets}
  \item{\code{ifa3g}}{Was given but never took it}
  \item{\code{ifa3h}}{Not interested}
  \item{\code{ifa3i}}{No reason}
  \item{\code{ifa4}}{Did you take IFA?}
  \item{\code{ifa4a}}{Reasons: Concerns about side effects}
  \item{\code{ifa4b}}{Reasons: Don't need it}
  \item{\code{ifa4c}}{Reasons: I was told not to take it}
  \item{\code{ifa4d}}{Reasons: I don't think it helps}
  \item{\code{ifa5}}{How many days did you take IFA?}
  \item{\code{ifa6}}{Took IFA for at least 90 days}
  \item{\code{ifa7}}{Took IFA for 180 days}
}}
\source{
Liberia Coverage Survey Round 2 - 27 August to 30 September 2019
}
\usage{
ifaInt
}
\description{
Interpolated iron-folic acid (IFA) indicator set Round 2
}
\keyword{datasets}
