% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-interpolate_indicators.R
\name{interpolate_indicators}
\alias{interpolate_indicators}
\title{Function to interpolate indicators from the Liberia Coverage Survey}
\usage{
interpolate_indicators(indicator = c("ifaDF", "iycfDF", "mnpDF", "vitDF",
  "anthroDF", "cmamDF"), core.columns = c("spid", "cid", "did", "eid",
  "motherID", "m2"), coords, hexgrid, idp = 2)
}
\arguments{
\item{indicator}{A character vector of indicator data.frame names}

\item{core.columns}{A vector of variable names included in indicator
data.frames}

\item{coords}{A data.frame containing per enumeration area centroid
coordinates}

\item{hexgrid}{A SpatialPoints class object containing locations of
interpolation}

\item{idp}{Inverse distance power. Default is 2.}
}
\value{
A data.frame as long as `hexgrid` containing
  interpolated indicator values at each location of `hexgrid`
}
\description{
Function to interpolate indicators from the Liberia Coverage Survey
}
\examples{
interpolate_indicators(indicator = "vitDF",
                       coords = sampleList.r2[ , c("EFEACODE", "lon", "lat")],
                       hexgrid = gmHexGrid)

}
