% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iycfInt}
\alias{iycfInt}
\title{Interpolated Infant and Young Child Feeding (IYCF) counselling coverage
indicator set Round 2}
\format{A data.frame with 10 columns:
\describe{
  \item{\code{icf1}}{Hear about IYCF counselling?}
  \item{\code{icf2}}{Attend IYCF counselling?}
  \item{\code{icf2a}}{Reasons: timing not convenient}
  \item{\code{icf2b}}{Reasons: Not interested}
  \item{\code{icf2c}}{Reasons: Do not trust counsellor}
  \item{\code{icf2d}}{Reasons: Don't think I need it}
  \item{\code{icf2e}}{Reasons: Did not go to hospital}
  \item{\code{icf2f}}{Reasons: No one told me about it/not invited}
  \item{\code{icf2g}}{Reasons: No time}
  \item{\code{icf3}}{Received/purchased IFA?}
}}
\source{
Liberia Coverage Survey Round 2 - 27 August to 30 September 2019
}
\usage{
iycfInt
}
\description{
Interpolated Infant and Young Child Feeding (IYCF) counselling coverage
indicator set Round 2
}
\keyword{datasets}
