% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liberiaData.R
\docType{package}
\name{liberiaData}
\alias{liberiaData}
\alias{liberiaData-package}
\title{Liberia Coverage Survey Data Handler}
\description{
The Liberia Coverage Survey for the assessment of CMAM coverage, 
    micronutrient powder supplementation coverage, vitamin A supplementation 
    coverage, iron folic acid supplementation coverage and IYCF counselling 
    coverage has been commissioned by UNICEF Liberia with funding from Power of 
    Nutrition. In support of this survey, this package for data manipulation,
    processing and analysis has been developed.
}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (0000-0002-4887-4415)

}
\keyword{internal}
