% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-merge_data.R
\name{merge_repeats}
\alias{merge_repeats}
\title{Function to merge main ODK form with repeats}
\usage{
merge_repeats(x, rep.name)
}
\arguments{
\item{x}{A list of data.frames produced by `get_liberia_data` after
retreiving data from ODK server.}

\item{rep.name}{A character value specifying name of repeat form to merge
with main form.}
}
\value{
A data.frame with rows equal to repeat data merged with main form
}
\description{
Function to merge main ODK form with repeats
}
\examples{
\dontrun{
merge_repeats(x = coverageData, rep.name = "child_repeat")
}

}
