% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnpDF}
\alias{mnpDF}
\title{Micronutrient Powder (MNP) supplementation coverage Round 2}
\format{A data.frame with 28 columns:

* `spid` - Sampling point identifier
* `cid` - County identifier. 1 = Urban Montserrado; 2 = Grand Bassa
* `did` - District identifier
* `eid` - Enumeration area identifier
* `motherID` - Mother identifier
* `m2` - Mother age
* `mnp1` - Heard about MNP?
* `mnp2` - Received/purchased MNP?
* `mnp2a` - Reasons: Health centre ran out
* `mnp2b` - Reasons: Took too long to get MNP
* `mnp2c` - Reasons: Too expensive
* `mnp2d` - Reasons: Child doesn't need it
* `mnp2e` - Reasons: Heard it doesn't work/help
* `mnp2f` - Reasons: Don't know/not seen/not heard of MNP
* `mnp2g` - Reasons: Have not gone to hospital
* `mnp2h` - Reasons: Have not received any from hospital/centre
* `mnp2i` - Reasons: Hospital didn't tell me about it
* `mnp3` - Take MNP?
* `mnp3a` - Reasons: Too expensive
* `mnp3b` - Reasons: Not available in the market
* `mnp3c` - Reasons: Do not need MNP
* `mnp3d` - Reasons: Heard other's bad experience
* `mnp3e` - Reasons: Advised not to use it
* `mnp3f` - Reasons: Not seen other mothers use it
* `mnp3g` - Reasons: Don't trust the product
* `mnp3h` - Reasons: Using another product
* `mnp3i` - Reasons: Haven't seen it
* `mnp3j` - Reasons: Child refused to take it
* `mnp4` - How many times child ate MNP in the past week?}
\source{
Liberia Coverage Survey Round 2 - 27 August to 30 September 2019
}
\usage{
mnpDF
}
\description{
Micronutrient Powder (MNP) supplementation coverage Round 2
}
\keyword{datasets}
