% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnpInt}
\alias{mnpInt}
\title{Interpolated Micronutrient Powder (MNP) supplementation coverage indicator
set Round 2}
\format{
A list of 2 data.frames with 23 columns:
\describe{
  \item{\code{mnp1}}{Heard about MNP?}
  \item{\code{mnp2}}{Received/purchased MNP?}
  \item{\code{mnp2a}}{Reasons: Health centre ran out}
  \item{\code{mnp2b}}{Reasons: Took too long to get MNP}
  \item{\code{mnp2c}}{Reasons: Too expensive}
  \item{\code{mnp2d}}{Reasons: Child doesn't need it}
  \item{\code{mnp2e}}{Reasons: Heard it doesn't work/help}
  \item{\code{mnp2f}}{Reasons: Don't know/not seen/not heard of MNP}
  \item{\code{mnp2g}}{Reasons: Have not gone to hospital}
  \item{\code{mnp2h}}{Reasons: Have not received any from hospital/centre}
  \item{\code{mnp2i}}{Reasons: Hospital didn't tell me about it}
  \item{\code{mnp3}}{Take MNP?}
  \item{\code{mnp3a}}{Reasons: Too expensive}
  \item{\code{mnp3b}}{Reasons: Not available in the market}
  \item{\code{mnp3c}}{Reasons: Do not need MNP}
  \item{\code{mnp3d}}{Reasons: Heard other's bad experience}
  \item{\code{mnp3e}}{Reasons: Advised not to use it}
  \item{\code{mnp3f}}{Reasons: Not seen other mothers use it}
  \item{\code{mnp3g}}{Reasons: Don't trust the product}
  \item{\code{mnp3h}}{Reasons: Using another product}
  \item{\code{mnp3i}}{Reasons: Haven't seen it}
  \item{\code{mnp3j}}{Reasons: Child refused to take it}
  \item{\code{mnp4}}{How many times child ate MNP in the past week?}
}
}
\source{
Liberia Coverage Survey Round 2 - 27 August to 30 September 2019
}
\usage{
mnpInt
}
\description{
Interpolated Micronutrient Powder (MNP) supplementation coverage indicator
set Round 2
}
\keyword{datasets}
