% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psuDataGB}
\alias{psuDataGB}
\title{Population dataset for enumeration areas in Grand Bassa}
\format{An object of class \code{data.frame} with 90 rows and 2 columns.}
\usage{
psuDataGB
}
\description{
\describe{
  \item{\code{psu}}{Primary sampling unit corresponding to the enumeration area code}
  \item{\code{pop}}{Population}
}
}
\keyword{datasets}
