% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psuDataGM}
\alias{psuDataGM}
\title{Population dataset for enumeration areas in Urban Montserrado}
\format{
An object of class \code{data.frame} with 88 rows and 2 columns.
}
\usage{
psuDataGM
}
\description{
\describe{
  \item{\code{psu}}{Primary sampling unit corresponding to the enumeration area code}
  \item{\code{pop}}{Population}
}
}
\keyword{datasets}
