% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-recode_anthro.R
\name{recode_age}
\alias{recode_age}
\title{Function to recode age variable and convert to age in days}
\usage{
recode_age(df)
}
\arguments{
\item{df}{A coverage data.frame collected for the Liberia Coverage Survey}
}
\value{
A vector of recoded ages based on any re-testing done
}
\description{
Function to recode age variable and convert to age in days
}
\examples{

recode_age(df = coverageData.r2)

}
