% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-recode_cmam.R
\name{recode_cmam}
\alias{recode_cmam}
\title{Function to recode CMAM coverage indicators}
\usage{
recode_cmam(df, core.columns = c("spid", "cid", "did", "eid", "motherID",
  "m2"))
}
\arguments{
\item{df}{A coverage data.frame collected for the Liberia Coverage Survey}

\item{core.columns}{A vector of variable names to include in resulting
data.frame}
}
\value{
A data.frame of recoded CMAM coverage indicators
}
\description{
Function to recode CMAM coverage indicators
}
\examples{
recode_cmam(df = coverageData.r2)

}
