% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-recode_anthro.R
\name{recode_height}
\alias{recode_height}
\title{Function to recode height variable}
\usage{
recode_height(df)
}
\arguments{
\item{df}{A coverage data.frame collected for the Liberia Coverage Survey}
}
\value{
A vector of recoded heights based on any re-testing done
}
\description{
Function to recode height variable
}
\examples{

recode_height(df = coverageData.r2)

}
