% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-create_ifa.R
\name{recode_ifa}
\alias{recode_ifa}
\title{Function to recode iron-folic acid coverage indicators}
\usage{
recode_ifa(df, core.columns = c("cid", "did", "eid", "motherID", "m2"))
}
\arguments{
\item{df}{A coverage data.frame collected for the Liberia Coverage Survey}

\item{core.columns}{A vector of variable names to include in resulting
data.frame}
}
\value{
A data.frame of iron-folic acid coverage indicators
}
\description{
Function to recode iron-folic acid coverage indicators
}
\examples{
recode_ifa(df = coverageData.r2)

}
