% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-recode_iycf.R
\name{recode_iycf}
\alias{recode_iycf}
\title{Function to recode IYCF counselling coverage indicators}
\usage{
recode_iycf(df, core.columns = c("spid", "cid", "did", "eid", "motherID",
  "m2"), ls = NULL)
}
\arguments{
\item{df}{A coverage data.frame collected for the Liberia Coverage Survey}

\item{core.columns}{A vector of variable names to include in resulting
data.frame}

\item{ls}{A list of thematically named string pattern vectors to use for
searching and organising `others` responses for icf2a}
}
\value{
A data.frame of IYCF counselling coverage indicators
}
\description{
Function to recode IYCF counselling coverage indicators
}
\examples{
icf2e <- c("delivery", "health", "hospital") # Did not go to hospital
icf2f <- c("Doctor", "invited", "talk", "counselling") # No one told me about it/not invited
icf2g <- c("time") # No time
ls <- list(icf2e, icf2f, icf2g)
names(ls) <- paste("icf2", letters[5:7], sep = "")
recode_iycf(df = coverageData1.r2, ls = ls)

}
