% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-recode_anthro.R
\name{recode_oedema}
\alias{recode_oedema}
\title{Function to recode oedema variable}
\usage{
recode_oedema(df)
}
\arguments{
\item{df}{A coverage data.frame collected for the Liberia Coverage Survey}
}
\value{
A vector of recoded oedema results based on any re-testing done
}
\description{
Function to recode oedema variable
}
\examples{

recode_oedema(df = coverageData.r2)

}
