% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-recode_others.R
\name{recode_others}
\alias{recode_others}
\title{Function to recode others responses}
\usage{
recode_others(df, var, ls, label = paste(var, 1:length(ls), sep = "."))
}
\arguments{
\item{df}{A data.frame containing data from the Liberia Coverage Survey}

\item{var}{A character value for variable name containing `others` responses}

\item{ls}{A list of thematically named string pattern vectors to use for
searching and organising `others` responses}

\item{label}{A character vector for names of the resulting data.frame.}
}
\value{
A data.frame with same number of rows as `df` and number of columns
  equal to the length of `ls` containing recoded values
}
\description{
Function to recode others responses
}
\examples{
ifa3d <- c("delivery", "hospital", "constantly") # Did not go to hospital
ifa3e <- c("know") # Don't know about the tablets
ifa3f <- c("any", "go") # Didn't get any tablets
ifa3g <- c("given") # Was given but never took it
ifa3h <- c("interest", "interested", "Not") # Not interested
ifa3i <- c("Reasons") # No reasons
ls1 <- list(ifa3d, ifa3e, ifa3f, ifa3g, ifa3h, ifa3i)
names(ls1) <- paste("ifa3", letters[4:9], sep = "")
recode_others(df = coverageData1.r2, var = "ifa3a_other", ls = ls1)

}
