% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-recode_screening.R
\name{recode_screening}
\alias{recode_screening}
\title{Function to recode nutrition screening coverage indicators}
\usage{
recode_screening(df, core.columns = c("spid", "cid", "did", "eid",
  "motherID", "m2"))
}
\arguments{
\item{df}{A coverage data.frame collected for the Liberia Coverage Survey}

\item{core.columns}{A vector of variable names to include in resulting
data.frame}
}
\value{
A data.frame of recoded nutrition screening coverage indicators
}
\description{
Function to recode nutrition screening coverage indicators
}
\examples{
recode_screening(df = coverageData.r2)

}
