% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-recode_vita.R
\name{recode_vita}
\alias{recode_vita}
\title{Function to recode vitamin A supplementation coverage indicators}
\usage{
recode_vita(df, core.columns = c("cid", "did", "eid", "motherID", "m2"),
  ls = NULL)
}
\arguments{
\item{df}{A coverage data.frame collected for the Liberia Coverage Survey}

\item{core.columns}{A vector of variable names to include in resulting
data.frame}

\item{ls}{A list of thematically named string pattern vectors to use for
searching and organising `others` responses for vit1a}
}
\value{
A data.frame of vitamin A supplementation coverage indicators
}
\description{
Function to recode vitamin A supplementation coverage indicators
}
\examples{
\dontrun{
ifa3d <- c("delivery", "hospital", "constantly") # Did not go to hospital
ifa3e <- c("know") # Don't know about the tablets
ifa3f <- c("any", "go") # Didn't get any tablets
ifa3g <- c("given") # Was given but never took it
ifa3h <- c("interest", "interested", "Not") # Not interested
ifa3i <- c("Reasons") # No reasons
ls1 <- list(ifa3d, ifa3e, ifa3f, ifa3g, ifa3h, ifa3i)
names(ls1) <- paste("ifa3", letters[4:9], sep = "")
recode_ifa(df = coverageData.r2, ls1 = ls1)
}

}
