% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-recode_anthro.R
\name{recode_weight}
\alias{recode_weight}
\title{Function to recode weight variable}
\usage{
recode_weight(df)
}
\arguments{
\item{df}{A coverage data.frame collected for the Liberia Coverage Survey}
}
\value{
A vector of recoded weights based on any re-testing done
}
\description{
Function to recode weight variable
}
\examples{

recode_weight(df = coverageData.r2)

}
