% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{screenDF}
\alias{screenDF}
\title{Nutrition screening coverage indicators}
\format{
A data.frame with 10 columns:
\describe{
  \item{\code{spid}}{Sampling point identifier}
  \item{\code{cid}}{County identifier. 1 = Urban Montserrado; 2 = Grand Bassa}
  \item{\code{did}}{District identifier}
  \item{\code{eid}}{Enumeration area identifier}
  \item{\code{motherID}}{Mother identifier}
  \item{\code{m2}}{Mother age}
  \item{\code{muac.screen}}{Has child been measured with a MUAC tape in past month?}
  \item{\code{oedema.screen}}{Has child been checked for oedema in the past month?}
  \item{\code{weight.screen}}{Has child's weight been measured in the past month?}
  \item{\code{height.screen}}{Has child's height been measured in the past month?}
}
}
\source{
Liberia Coverage Survey Round 2 - 27 August to 30 September 2019
}
\usage{
screenDF
}
\description{
Nutrition screening coverage indicators
}
\keyword{datasets}
