% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{villageData.r2}
\alias{villageData.r2}
\title{Liberia Coverage Survey Village Data Round 2}
\format{An object of class \code{data.frame} with 90 rows and 11 columns.}
\source{
Liberia Coverage Survey Round 2 - 27 August to 30 September 2019
}
\usage{
villageData.r2
}
\description{
Liberia Coverage Survey Village Data Round 2
}
\keyword{datasets}
