% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{villageData2}
\alias{villageData2}
\title{Liberia Coverage Survey Village Data Round 2}
\format{A data.frame with 11 columns

* `cid` - County identifier
* `county` - County name
* `did` - District identifier
* `eid` - Enumerator identifier
* `lat` - Provided latitude coordinate
* `lon` - Provided longitude coordinate
* `pop` - Population
* `coords.Latitude` - Recorded latitude coordinate
* `coords.Longitude` - Recorded longitude coordinate
* `coords.Altitude` - Recorded altitude
* `coords.Accuracy` - Accuracy of coordinates recording}
\source{
Liberia Coverage Survey Round 2 - 27 August to 30 September 2019
}
\usage{
villageData2
}
\description{
Liberia Coverage Survey Village Data Round 2
}
\keyword{datasets}
