% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vitBoot}
\alias{vitBoot}
\title{Boostrap results for vitamin A indicators}
\format{
A data.frame with 11 rows and 7 columns
\describe{
  \item{\code{county}}{County name}
  \item{\code{varLabel}}{Short code of indicators}
  \item{\code{df}}{Name of data.frame to which indicator belongs}
  \item{\code{varNames}}{Long names of indicators}
  \item{\code{estimate}}{Indicator estimate}
  \item{\code{lcl}}{Indiator 95\% lower confidence limits}
  \item{\code{ucl}}{Indiator 95\% lower confidence limits}
}
}
\usage{
vitBoot
}
\description{
Boostrap results for vitamin A indicators
}
\keyword{datasets}
