% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vitInt}
\alias{vitInt}
\title{Interpolated Vitamin A supplementation coverage indicator set Round 2}
\format{A list of 2 data.frames with 11 columns:
\describe{
  \item{\code{vit1}}{Received vitamin A?}
  \item{\code{vit1a}}{Reasons: Health centre ran out}
  \item{\code{vit1b}}{Reasons: Took too long to get drops}
  \item{\code{vit1c}}{Reasons: Child doesn't need it}
  \item{\code{vit1d}}{Reasons: Heard it doesn't work/help}
  \item{\code{vit1e}}{Reasons: Child too young during latest campaign}
  \item{\code{vit1f}}{Reasons: Not around during last campaign}
  \item{\code{vit1g}}{Reasons: Didn't hear/know about it}
  \item{\code{vit1h}}{Reasons: Never been to hospital}
  \item{\code{vit1i}}{Reasons: Vaccine team didn't come/reach}
  \item{\code{vit1j}}{Reasons: Not interested}
}}
\source{
Liberia Coverage Survey Round 2 - 27 August to 30 September 2019
}
\usage{
vitInt
}
\description{
Interpolated Vitamin A supplementation coverage indicator set Round 2
}
\keyword{datasets}
