% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anthroDF}
\alias{anthroDF}
\title{Child anthropometry indicators Round 2}
\format{An object of class \code{data.frame} with 3425 rows and 29 columns.}
\source{
Liberia Coverage Survey Round 2 - 27 August to 30 September 2019
}
\usage{
anthroDF
}
\description{
Child anthropometry indicators Round 2
}
\keyword{datasets}
