% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmamDF}
\alias{cmamDF}
\title{Community-based Management of Acute Malnutrition (CMAM) coverage indicators
Round 2}
\format{A data.frame with 35 columns:

* `cid` - County identifier. 1 = Urban Montserrado; 2 = Grand Bassa
* `did` - District identifier
* `eid` - Enumeration area identifier
* `motherID` - Mother identifier
* `m2` - Mother age
* `cin` - SAM case in programme
* `cout` - SAM case not in programme
* `rin` - Recovering case in programme
* `q1` - Is child sick?
* `q1a` - Child suffering from: malnutrition
* `q1b` - Child suffering from: diarrhoea
* `q1c` - Child suffering from: fever
* `q1d` - Child suffering from: cough
* `q1e` - Child suffering from: malaria
* `q1f` - Child suffering from: don't know
* `q2` - Do you think your child is too small?
* `q3` - Know of treatment
* `q4a` - Came to health centre but was not measured
* `q4b` - Too far
* `q4c` - No time/too busy
* `q4d` - Mother/carer is sick
* `q4e` - Cannot carry more than one child
* `q4f` - Feels ashamed about coming
* `q4g` - Security problems
* `q4h` - No one else to take care of other siblings
* `q4i` - Rejected by programme
* `q4j` - Other parent's children have been rejected
* `q4k` - Husband refused
* `q4l` - Thought necessary to be enrolled in hospital first
* `q4m` - Does not think programme can help children
* `q4n` - Got diarrhoea
* `q4o` - Ration has worms
* `q4p` - Out of stock
* `q4q` - Lack of money to pay for transport
* `q4r` - SAM case in SFP}
\source{
Liberia Coverage Survey Round 2 - 27 August to 30 September 2019
}
\usage{
cmamDF
}
\description{
Community-based Management of Acute Malnutrition (CMAM) coverage indicators
Round 2
}
\keyword{datasets}
