% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codebook}
\alias{codebook}
\title{Liberia Coverage Survey data codebook}
\format{A tibble with 3 columns and 152 rows:
\describe{
  \item{\code{variable}}{Variable name}
  \item{\code{question}}{Question or data assigned to \code{variable}}
  \item{\code{choices}}{Possible responses for specific \code{variable}}
 }}
\source{
Liberia Coverage Survey Round 2 - 27 August to 30 September 2019
}
\usage{
codebook
}
\description{
Liberia Coverage Survey data codebook
}
\keyword{datasets}
