% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-create_sp.R
\name{create_sp}
\alias{create_sp}
\title{Create sp class versions of datasets}
\usage{
create_sp(df, coords, proj)
}
\arguments{
\item{df}{A data.frame of recoded indicator values for the Liberia Coverage
Survey}

\item{coords}{A data.frame containing per enumeration area centroid
coordinates}

\item{proj}{A projection string of class CRS}
}
\value{
An `sp` class version of datasets
}
\description{
Create sp class versions of datasets
}
\examples{
create_sp(df = coverageData.r2, coords = sampleList.r2,
          proj = sp::CRS(sp::proj4string(gmHexGrid)))

}
