% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iycfDF}
\alias{iycfDF}
\title{Infant and Young Child Feeding (IYCF) counselling coverage Round 2}
\format{A data.frame with 15 columns:

* `cid` - County identifier. 1 = Urban Montserrado; 2 = Grand Bassa
* `did` - District identifier
* `eid` - Enumeration area identifier
* `motherID` - Mother identifier
* `m2` - Mother age
* `icf1` - Hear about IYCF counselling?
* `icf2` - Attend IYCF counselling?
* `icf2a` - Reasons: timing not convenient
* `icf2b` - Reasons: Not interested
* `icf2c` - Reasons: Do not trust counsellor
* `icf2d` - Reasons: Don't think I need it
* `icf2e` - Reasons: Did not go to hospital
* `icf2f` - Reasons: No one told me about it/not invited
* `icf2g` - Reasons: No time
* `icf3` - Received/purchased IFA?}
\source{
Liberia Coverage Survey Round 2 - 27 August to 30 September 2019
}
\usage{
iycfDF
}
\description{
Infant and Young Child Feeding (IYCF) counselling coverage Round 2
}
\keyword{datasets}
