% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-recode_others.R
\name{recode_others}
\alias{recode_others}
\title{Function to recode others responses}
\usage{
recode_others(df, var, ls)
}
\arguments{
\item{df}{A data.frame containing data from the Liberia Coverage Survey}

\item{var}{A character value for variable name containing `others` responses}

\item{ls}{A list of thematically named string pattern vectors to use for
searching and organising `others` responses}
}
\value{
A data.frame with same number of rows as `df` and number of columns
  equal to the length of `ls` containing recoded values
}
\description{
Function to recode others responses
}
\examples{
\dontrun{
recode_others(df = coverageData2, var = "ifa3a_other", ls = )
}

}
