% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{villageData.r2}
\alias{villageData.r2}
\title{Liberia Coverage Survey Village Data Round 2}
\format{A data.frame with 11 columns
\describe{
  \item{\code{cid}}{County identifier}
  \item{\code{county}}{County name}
  \item{\code{did}}{District identifier}
  \item{\code{eid}}{Enumerator identifier}
  \item{\code{lat}}{Provided latitude coordinate}
  \item{\code{lon}}{Provided longitude coordinate}
  \item{\code{pop}}{Population}
  \item{\code{coords.Latitude}}{Recorded latitude coordinate}
  \item{\code{coords.Longitude}}{Recorded longitude coordinate}
  \item{\code{coords.Altitude}}{Recorded altitude}
  \item{\code{coords.Accuracy}}{Accuracy of coordinates recording}
}}
\source{
Liberia Coverage Survey Round 2 - 27 August to 30 September 2019
}
\usage{
villageData.r2
}
\description{
Liberia Coverage Survey Village Data Round 2
}
\keyword{datasets}
